pyproject.toml
[project]
name = "ght-qcxr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mlx-embeddings>=0.0.5",
    "mlx-vlm>=0.3.9",
    "numpy>=2.4.2",
    "pillow>=12.1.0",
    "pydicom>=3.0.1",
    "rich>=14.3.2",
    "transformers>=5.0.0",
]
